infra/pyproject.toml
[project]
name = "hf-ml-platform-auth-infra"
version = "0.1.0"
description = "ECS infrastructure for the HuggingFace ML Inference Platform auth service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.180.0",
    "constructs>=10.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[dependency-groups]
dev = [
    "tox>=4.0.0",
    "ruff>=0.9.0",
]
